ファイルの切り分け、許可ヘッダの追加
